src/backend/requirements.txt

fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dateutil>=2.8.2
